Kill logcat with SEGV so we get a tombstone on timeout.
Looking at the history of this code, we've had hard to reproduce issues
here since at least 2016. The most recent bug doesn't even have any
information to let us work out whether the problem is with logcat, with
timeout, or with the Java code calling them. Let's use SIGSEGV so we'll
at least have a tombstone if timeout has to kill logcat.

Bug: http://b/187005648
Test: treehugger
Change-Id: I224555c631d784647dafc30641c2c8a447884971