Move mForceDeniedNapps check into synchronized block
This fixes a race condition that existed when using the test command
that allows force denying access to a nanoapp. This race condition
occurred when updateNanoAppAuthState was called immediately following
the force denial command invocation which meant that it was waiting on
the lock to be released after the force denial went through and was
using stale mForceDeniedNapps which caused the authorization to be
granted again.

Fixes: 187437820
Test: atest PtsChreTestCases
Change-Id: I39fc05abc95f3d7fe0c929ddce9b212f729506c5