Fix lockscreen notification + media counting.
Background: A regression was introduced when something about the MediaHeaderView's visibility was changed, which caused that view to be counted when showing notifications but not when counting how many should be shown.  That regression caused b/188000471 which was fixed by ag/I93b4b6039eb812619787744f893fe1924e1afc4d.  However, that fix only addressed half the problem, as it resulted in the media player always being shown, regardless of the hard-coded notification max.

This change brings the logic of these two methods in line with each other, so that the number of notifications shown *includes* the media player, but ensures that a zero-heignt media player is not treated as a notification (similar to how any GONE views are ignored).

Fixes: 189102584
Test: manual
Change-Id: Ic7447941593f08ed504f3870af5a72d2628a043e