Tidy configuration associated with geotz
Tidy configuration associated with geolocation time zone detection.

This commit contains a number of documentation fixes.

It also is part of a topic that moves the configuration of the secondary
location time zone provider to a Google-specific overlay.

Note: There will be a follow up later in the release cycle to remove the
hosting of the secondary LTZP in the system server.

Bug: 178169250
Bug: 174027180
Test: treehugger
Change-Id: Icb56bb78e06fe6ac30fd5671208994aebcff29b8