Don't cache IME surface in fullscreen mode
Dont cache IME surface when IME was in fullscreen mode. This is done in
order to fix IME closing when it is used with RecyclerView. There can be
a special case where RecyclerView detaches the view holding mServedView
when IME is in fullscreen mode

While exact reason is still a mystery, short term solution is to not
cache IME when it was in fullscreen mode.

Fix: 187772544
Bug: 188818557
Bug: 167948123

Test: Manually using steps in bug
Change-Id: I1194d08a00622f1dfa232209a70dcb0797ba192b