Keep wallpaper as a special-case in transitions
Wallpaper unfortunately needs to be handled specially by
transition players because of it's arrangement in the hierarchy
(being in a separate root displayareay) and because it has
specific ordering mechanics during animations.

Now that DisplayAreas are "organized", they interfere with
wallpaper's handling as a special-case in a transition. To prevent
this, explicitly prevent any of wallpaper's ancestor displayareas
from being part of the transition.

Bug: 187461719
Test: launch a landscape app from portrait launcher. Observe
      wallpaper not disappearing.
Change-Id: Id9e86f14d380b5ed8050b303edf5f1b48875ff20