Always log.d when restoring settings.
This commit removes the `if (DEBUG)` wrapper so that we always debug log
for settings that are restored by the backup agent.

This will help investigating bug reports about some settings not getting
restored, as we will then see the list of successfully restored settings
in the bugreport.   Restore is an infrequent operation so this doesn't
add much noise to logs.

Bug: 161106112
Change-Id: I0f07474fe211f5fcc6cf89ce701378265057b8f5
Test: atest com.android.providers.settings.SettingsBackupAgentTest