Remove call to set wakeLock WorkSource from VibrationThread constructor
A new  VibrationThread might be created and dropped if the ongoing
thread takes long to be finished and a new vibration arrived to the
service.

The wakelock is only acquired when a thread starts, so move the calls to
configure the worksource to the run() method.

Fix: 193742818
Test: manual
Change-Id: I7d7953ec9e065bcec78110b44daf3ef57693164d
(cherry picked from commit c673fad8bf8cc6e17e05d8268ebffc7e60af26d7)
Merged-In: I7d7953ec9e065bcec78110b44daf3ef57693164d