Fix empty side stage after dropping foreground app icon to split
To prevent it entering stage split with empty side stage root, update to
not activate split explicitly when dropping app icons to split. Only
active stage split when the target activity is actually launching in
the side stage root.

Fix: 193200163
Test: drap and drop app icons to enter split, observed it won't active
stage split if the app is already showing in foreground.

Change-Id: I1cad96f6265eb1a14369bfc708871f701764066a